Create useQuery hook, fix bug with delayed/real-time price in title byline, fix bug with escaped text showing in news feed
